Fix all bugs for lighty-yang-validator reported by SonarCloud.
Fixed 2 bugs where try-with-resources or close stream in finally clause  was recommended to use.
Fixed 24 bugs where presence of value in Optional<> variable was not checked before accessing the value.
Refractored some if statements where presence of value in Optional<> variable was checked before but sonarcloud reported it as bug.