Fix JsonTree issue for multiple yang files
The issue with Format JsonTree is that it creates an output file that contains multiple
JSON objects, one for each tested module. The results is an invalid JSON file, as the
objects are not wrapped in a single outer object. The desired output for each tested
module is to be wrapped in an array that is itself contained within a single outer object.

Assertion test failures expected, since output of JsonTree has changed.

Signed-off-by: tobias.pobocik <tobias.pobocik@pantheon.tech>