Update decimal-data value from 5 to 3
The decimal-data leaf in the YANG model ietf-interfaces-modified
is bounded to the range "1 .. 3.14 | 10 | 20..max".Prior to this commit,
the value of decimal-data was set to 5, which is outside the allowed range.
This commit updates the value of decimal-data to 3, which is within the allowed range.

This issue became valid after the addition of range constraints in newer version of yangtools:
https://github.com/opendaylight/yangtools/commit/293d2ff97740f582cd402e3b4850d22d6752b9db

JIRA:LIGHTY-151
Signed-off-by: tobias.pobocik <tobias.pobocik@pantheon.tech>