Improve testing for multiple modules
Previously, there was no test coverage for the output of multiple
modules. To address this, this commit adds a new test that parses
two modules and compares their combined output against the expected
output. This provides more comprehensive testing and ensures that
the json-tree modules are correctly processing multiple inputs.

JIRA: LIGHTY-229
Signed-off-by: tobias.pobocik <tobias.pobocik@pantheon.tech>
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>